Fix handling of params in `removeFixedJoints`.
Now, if the parent or child link of a fixed joint has parameterized
inertia, mass, or com properties, `removeFixedJoints` will attempt to
parameterized the corresponding properties of the combined link. This
can fail fairly easily when updating the com, as, for links with
non-zero mass, this requires dividing by the sum of the link masses.
Despite this limitations, the current implementation is quite useful in
that it allows for the construction of `PlanarRigidBodyManipulator`
objects whose floating-base links have inertial parameters.