Handle neg. params in `lumpedToOriginalParams`
This uses assumptions that are specific to static ID. It assumes
that each  monomial is the product containing at most one mass
(non-negative) and at most one displacement (free). Moreover, it assumes
that each displacement appears in only one monomial. Under these
assumptions, if the estimate of a given monomial is negative, the
displacement parameter present in that monomial must also be negative.
This allows us to change to a set of variables which are all
non-negative and then change back after solving the problem.