Add repro test for Eigen/clang weirdness
The new test (which fails) reproduces behavior observed in FCL's box-box
intersection algorithm. Elements of a Matrix3d copied from an input
argument contain garbage, when certain other conditions are met. I've
reduced the example as far as I can without loosing the weird behavior.
The issue does NOT occur if the test is compiled with gcc-5 or with the
optimization level set to -O1.