[solvers] Add L2NormCost isa Cost
Towards #15366.

Note: I've added the python binding for the class, but we currently don't have any AddCost variants that will support adding it to a MathematicalProgram.  I looked at that, but it's more than I want to bite off now.  As is, this binding will still work for my shortest path problem code.